JP-5692 - Crash on reading quality uniqueness constraint when adding interval data validator and storing flagged data over existing (unflagged) data

+++ coko/com.elster.jupiter.validators/src/main/java/com/elster/jupiter/validators/impl/MissingValuesValidator.java
@@ -30,7 +30,7 @@ class MissingValuesValidator extends AbstractValidator {
     private static final String READING_QUALITY_TYPE_CODE = "3.5.259";
     
     private Interval interval;
-    private int millisBetweenReadings;
+    private long millisBetweenReadings;
     private BitSet bitSet;
     private int expectedReadings;
     private boolean openEnded;
